Clippy fix: Move `SoyBoy` to mod.rs
